refactor: split index.ts into setupMiddleware, registerRoutes, initServices
